requirements.txt
Flask==0.10.1
Flask-Bootstrap==3.0.2.3
Flask-Login==0.2.7
Flask-SQLAlchemy==1.0
Flask-WTF==0.9.3
Flask-Mail==0.9.1
Jinja2==2.7.1
MarkupSafe==0.18
MySQL-python==1.2.4
PyRSS2Gen==1.0.0
SQLAlchemy==0.8.4
WTForms==1.0.5
Werkzeug==0.9.4
gunicorn==18.0
itsdangerous==0.23
python-dateutil==1.5
virtualenv==1.10.1
psycopg2==2.5.1
pyOpenSSL==0.13
requests==2.1.0
pytz==2012d
sqlalchemy-migrate==0.10.0
celery==3.1.17
